requirements.txt
matplotlib==3.8.2
numpy==1.26.4
pandas==2.2.2
Pillow
scikit_learn==1.3.2
seaborn==0.13.2
timm==0.9.12
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
